[IDLE-141] ViewModel을 참조하도록 수정
