+++ project/Projects/App/Sources/RootCoordinator/Main/Center /TabCoordinator/CenterProfileRegisterCoordinator.swift
