+++ project/Projects/Presentation/Feature/Center/Sources/View/RecruitmentPost/Board/CenterRecruitmentPostBoardVM.swift
@@ -12,6 +12,7 @@ import PresentationCore
 import RxCocoa
 import RxSwift
 import Entity
+import DSKit
 
 public protocol CenterRecruitmentPostBoardViewModelable: OnGoingPostViewModelable & ClosedPostViewModelable {
     var alert: Driver<DefaultAlertContentVO>? { get }
@@ -71,5 +72,33 @@ public class CenterRecruitmentPostBoardVM: CenterRecruitmentPostBoardViewModelab
     
     func publishClosedPostMocks() -> Single<Result<[CenterEmployCardVO], RecruitmentPostError>> {
         return .just(.success((0...10).map { _ in CenterEmployCardVO.mock }))
+    }
+}
+
+// MARK: 카드 뷰에 사용될 ViewModel
+class CenterEmployCardVM: CenterEmployCardViewModelable {
+    
+    // Init
+    var id: String
+    
+    // Output
+    var renderObject: Driver<CenterEmployCardRO>?
+    
+    // Input
+    var cardClicked: PublishRelay<Void> = .init()
+    var checkApplicantBtnClicked: PublishRelay<Void> = .init()
+    var editPostBtnClicked: PublishRelay<Void> = .init()
+    var terminatePostBtnClicked: PublishRelay<Void> = .init()
+    
+    init(vo: CenterEmployCardVO) {
+        self.id = vo.postId
+        
+        // MARK: RenderObject
+        let publishRelay: BehaviorRelay<CenterEmployCardRO> = .init(value: .mock)
+        renderObject = publishRelay.asDriver(onErrorJustReturn: .mock)
+        
+        publishRelay.accept(CenterEmployCardRO.create(vo))
+        
+        // MARK: 버튼 처리
     }
 }
